% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_correction_bh.R
\name{get_alphas_bh}
\alias{get_alphas_bh}
\title{Calculate Benjamini-Hochberg alphas for a list of p-values.}
\usage{
get_alphas_bh(
  p_values,
  Q = 0.05,
  output = "both",
  include_is_significant_column = TRUE
)
}
\arguments{
\item{p_values}{A numeric vector of p-values.}

\item{Q}{The false discovery rate (default 0.05).}

\item{output}{"print", "data_frame", or "both".}

\item{include_is_significant_column}{Boolean to include significance column.}
}
\value{
A data frame containing p-values, alphas, and significance.
}
\description{
Calculate Benjamini-Hochberg alphas for a list of p-values.
}
